Extract JSON conversion tests to another file
